+++ R/SNP.R
@@ -25,11 +25,13 @@ SNPinfo <- R6::R6Class(
     #' 3 columns:
     #' \itemize{
     #'  \item{\code{chr}:} {Chromosome holding the SNP}
-    #'  \item{\code{pos}:} {SNP position on the chromosome}
+    #'  \item{\code{physPos}:} {SNP physical position on the chromosome}
+    #'  \item{\code{linkMapPos}:} {SNP linkage map position on the chromosome}
     #'  \item{\code{SNPid}:} {SNP's IDs}
     #'  }
     SNPcoord = data.frame(chr = c(),
-                          pos = c(),
+                          physPos = c(),
+                          linkMapPos = c(),
                           SNPid = c(),
                           stringsAsFactors = FALSE),
 
@@ -48,7 +50,8 @@ SNPinfo <- R6::R6Class(
     #' @param SNPcoord [data.frame] Coordinate of all SNPs.
     #'
     #'   3 columns: \itemize{ \item{\code{chr}:} {Chromosome holding the SNP}
-    #'   \item{\code{pos}:} {SNP position on the chromosome}
+    #'   \item{\code{physPos}:} {SNP physical position on the chromosome}
+    #'   \item{\code{linkMapPos}:} {SNP linkage map position on the chromosome}
     #'   \item{\code{SNPid}:} {SNP's IDs} }
     #' @param specie [specie class] Specie of the SNPs
     #'   (see:\link[breedSimulatR]{specie})
@@ -57,15 +60,17 @@ SNPinfo <- R6::R6Class(
     #' # create specie
     #' mySpec <- specie$new(nChr = 3,
     #'                      lchr = c(100, 150, 200),
+    #'                      lchrCm = 100,
     #'                      verbose = FALSE)
     #'
     #' # simulate SNP
     #' SNPcoord <- data.frame(chr = c(rep("Chr1", 3),
     #'                                rep("Chr2", 4),
     #'                                rep("Chr3", 5)),
-    #'                        pos = c(sample(100, 3),
+    #'                        physPos = c(sample(100, 3),
     #'                                sample(150, 4),
     #'                                sample(200, 5)),
+    #'                        linkMapPos = NA,
     #'                        SNPid = sprintf(fmt = paste0("SNP%0", 2,"i"),
     #'                                        1:(3 + 4 + 5)))
     #'
@@ -82,9 +87,23 @@ SNPinfo <- R6::R6Class(
         stop('"class(SNPcoord)" must be "data.frame"')
       }
       # SNPcoord's colnames
-      if (!all(colnames(SNPcoord) %in% c("chr", "pos", "SNPid"))) {
-        stop('"colnames(SNPcoord)" must be "chr", "pos" and "SNPid"')
+      if (ncol(SNPcoord) == 3) {
+        if (!all(c("chr", "physPos", "SNPid") %in% colnames(SNPcoord))) {
+          stop('"colnames(SNPcoord)" must include "chr", "physPos" and "SNPid", and optionally "linkMapPos"')
+        }
+        # add linkMapPos column
+        SNPcoord$linkMapPos <- NA
+
+      } else if (ncol(SNPcoord) == 4) {
+        if (!all(c("chr", "physPos", "SNPid", "linkMapPos") %in% colnames(SNPcoord))) {
+          stop('"colnames(SNPcoord)" must include "chr", "physPos" and "SNPid", and optionally "linkMapPos"')
+        }
+        if (sum(is.na(SNPcoord$linkMapPos)) != 0
+            & sum(is.na(SNPcoord$linkMapPos)) != nrow(SNPcoord)) {
+          stop('"SNPcoord$linkMapPos" should either contains only NAs or none')
+        }
       }
+
       if (!all(unique(SNPcoord$chr) %in% specie$chrNames)) {
         stop(paste0('"Chromosomes\'names specified in "SNPcoord" do ',
                     'not match those specified in "specie"\n',
@@ -102,17 +121,30 @@ SNPinfo <- R6::R6Class(
         lapply(SNPcoord[, vapply(SNPcoord, is.factor, TRUE)], as.character)
 
       # convert to integer
-      SNPcoord$pos <- as.integer(SNPcoord$pos)
+      SNPcoord$physPos <- as.integer(SNPcoord$physPos)
 
       # sort position in increasing order (needed for the function
       # "findInterval" in individual's generateGametes method)
-      SNPcoord <- SNPcoord[order(SNPcoord$pos),]
+      SNPcoord <- SNPcoord[order(SNPcoord$physPos),]
 
 
       self$SNPcoord <- SNPcoord
       self$specie <- specie
       self$SNPcoordList <- split(SNPcoord,
                                  SNPcoord$chr)
+
+      # check SNP unicity and linkage map position order
+      lapply(self$SNPcoordList, function(subSNPcoord){
+        if (!identical(unique(subSNPcoord$physPos), subSNPcoord$physPos)){
+          stop("Some SNPs have the same physical position.")
+        }
+        # check the linkMapPos are also sorted like the physical position
+        if (is.unsorted(subSNPcoord$linkMapPos, na.rm = TRUE, strictly = TRUE)) {
+          stop("SNP's position order should be similar when sorted by physical position and by linkage map position.")
+        }
+        subSNPcoord
+      })
+
 
       self$ids <- lapply(specie$chrNames, function(chr){
         SNPcoord[SNPcoord$chr == chr, "SNPid"]
@@ -179,7 +211,7 @@ SNPinfo <- R6::R6Class(
 
       plotly::plot_ly(data = self$SNPcoord,
                       x = ~chr,
-                      y = ~pos,
+                      y = ~physPos,
                       type = "scatter",
                       mode = "markers",
                       alpha = alpha,