+++ R/phenotype.R
@@ -52,9 +52,9 @@ trait <- R6::R6Class(
     #' @examples
     #' mySpec <- specie$new(nChr = 10,
     #'                      lchr = 10^6,
+    #'                      lchrCm = 100,
     #'                      specName = "Geneticae Exempli",
-    #'                      ploidy = 2,
-    #'                      recombRate = 3/10^6)
+    #'                      ploidy = 2)
     #' SNPs <- SNPinfo$new(SNPcoord = exampleData$snpCoord,
     #'                     specie = mySpec)
     #'
@@ -206,9 +206,9 @@ phenotyper <- R6::R6Class(
     #' @examples
     #' mySpec <- specie$new(nChr = 10,
     #'                     lchr = 10^6,
+    #'                     lchrCm = 100,
     #'                     specName = "Geneticae Exempli",
-    #'                     ploidy = 2,
-    #'                     recombRate = 3/10^6)
+    #'                     ploidy = 2)
     #' SNPs <- SNPinfo$new(SNPcoord = exampleData$snpCoord,
     #'                    specie = mySpec)
     #' example_pop <- createPop(geno = exampleData$genotypes,