+++ tests/testthat/test_1-specie.R
@@ -13,10 +13,10 @@ test_that("specie initialization", {
 
   expect_error({mySpec <- specie$new(nChr = 3,
                                      lchr = c(100, 150, 200),
+                                     lchrCm = c(101, 151, 201),
                                      specName = "Geneticae Exempulus",
                                      ploidy = 2,
                                      mutRate = 10^-8,
-                                     recombRate = 10^-7,
                                      chrNames = c("C1", "C2", "C3"),
                                      verbose = F)},
                NA)
@@ -27,19 +27,22 @@ test_that("specie initialization", {
   expect_identical(mySpec$ploidy, 2)
   expect_identical(as.numeric(mySpec$lchr), c(100, 150, 200))
   expect_identical(names(mySpec$lchr), c("C1", "C2", "C3"))
+  expect_identical(as.numeric(mySpec$lchrCm), c(101, 151, 201))
+  expect_identical(names(mySpec$lchrCm), c("C1", "C2", "C3"))
   expect_identical(mySpec$mutRate, 10^-8)
   expect_identical(mySpec$chrNames, c("C1", "C2", "C3"))
-  expect_identical(mySpec$recombRate, 10^-7)
-  expect_output(specie$new(1, 10, verbose = T), paste("A new species has",
+  expect_output(specie$new(1, 10, 100, verbose = T), paste("A new species has",
                                                       "emerged: Undefinded !"))
-  expect_output(specie$new(1, 10, verbose = F), NA)
+  expect_output(specie$new(1, 10, 100, verbose = F), NA)
 
   expect_error(specie$new(nChr = 3.5,
                           lchr = c(100, 150, 200),
+                          lchrCm = c(101, 151, 201),
                           verbose = F),
                "nChr must be integer.")
   expect_error(specie$new(nChr = 3,
                           lchr = c(100, 150.2, 200),
+                          lchrCm = c(101, 151, 201),
                           verbose = F),
                "lchr must be integers.")
 })
@@ -49,11 +52,13 @@ test_that("specie initialization without optional values", {
 
   expect_error(specie$new(nChr = 3,
                           lchr = c(100, 150, 200),
+                          lchrCm = c(101, 151, 201),
                           verbose = F),
                NA)
 
   mySpec <- specie$new(nChr = 3,
                        lchr = c(100, 150, 200),
+                       lchrCm = c(101, 151, 201),
                        verbose = F)
 
   expect_identical(mySpec$specName, "Undefinded")
@@ -61,15 +66,17 @@ test_that("specie initialization without optional values", {
   expect_identical(mySpec$ploidy, 2)
   expect_identical(as.numeric(mySpec$lchr), c(100, 150, 200))
   expect_identical(names(mySpec$lchr), c("Chr1", "Chr2", "Chr3"))
+  expect_identical(as.numeric(mySpec$lchrCm), c(101, 151, 201))
+  expect_identical(names(mySpec$lchrCm), c("Chr1", "Chr2", "Chr3"))
   expect_identical(mySpec$mutRate, NA)
   expect_identical(mySpec$chrNames, c("Chr1", "Chr2", "Chr3"))
-  expect_identical(mySpec$recombRate, NA)
 })
 
 
 test_that("specie's \"getChrLength\" methods", {
   mySpec <- specie$new(nChr = 3,
                        lchr = c(100, 150, 200),
+                       lchrCm = c(101, 151, 201),
                        verbose = F)
 
   expect_is(mySpec$getChrLength(), "numeric")
@@ -82,20 +89,32 @@ test_that("specie's \"getChrLength\" methods", {
   expect_identical(names(mySpec$getChrLength()), c("Chr1", "Chr2", "Chr3"))
   expect_identical(names(mySpec$getChrLength(3)), "Chr3")
   expect_identical(names(mySpec$getChrLength("Chr1")), "Chr1")
+
+  mySpec <- specie$new(nChr = 3,
+                       lchr = c(100, 150, 200),
+                       lchrCm = c(101, 151, 201),
+                       verbose = F,
+                       chrNames = c("chr1","chr2","X"))
+  expect_identical(as.numeric(mySpec$getChrLength("X")), 200)
+
+
   })
+
+
 
 
 test_that("specie's \"print\" methods", {
   mySpec <- specie$new(nChr = 3,
                        lchr = c(100, 150, 200),
+                       lchrCm = c(101, 151, 201),
                        verbose = F)
 
-  expect_output(print(mySpec), paste("Name: Undefinded\\nNumber of",
-                                     "Chromosomes: 3\\nPloidy: 2\\nMutation",
-                                     "rate : NA\\nRecombination Rate:",
-                                     "NA\\nChromosome length:\\n",
-                                     "    chrNames chrLength\\nChr1",
-                                     "    Chr1       100\\nChr2",
-                                     "    Chr2       150\\nChr3",
-                                     "    Chr3       200"))
+  expect_output(print(mySpec), paste0("Name: Undefinded\\n",
+                                      "Number of Chromosomes: 3\\n",
+                                      "Ploidy: 2\\n",
+                                      "Chromosome length:\\n",
+                                      "     chrNames chrLength chrLengthCm\\n",
+                                      "Chr1     Chr1       100         101\\n",
+                                      "Chr2     Chr2       150         151\\n",
+                                      "Chr3     Chr3       200         201"))
 })